Move commands into commands module
