+++ crates/ubrn_cli/src/commands/mod.rs
@@ -3,10 +3,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/
  */
-pub(crate) mod codegen;
-pub(crate) mod commands;
-pub(crate) mod config;
 
-pub(crate) use codegen::get_files;
-pub(crate) use commands::AndroidArgs;
-pub(crate) use config::AndroidConfig;
+pub(crate) mod building;
+pub(crate) mod checkout;
+pub(crate) mod generate;
+
+pub(crate) use building::BuildArgs;
+pub(crate) use checkout::CheckoutArgs;
+pub(crate) use generate::GenerateArgs;
