+++ crates/ubrn_cli/src/jsi/android/commands.rs
@@ -19,10 +19,9 @@ use uniffi_bindgen::{
 };
 
 use crate::{
-    building::CommonBuildArgs,
-    config::{ExtraArgs, ProjectConfig},
+    commands::building::CommonBuildArgs,
+    config::{rust_crate::CrateConfig, ExtraArgs, ProjectConfig},
     jsi::android::config::Target,
-    rust::CrateConfig,
 };
 
 #[derive(Args, Debug)]
